renamed `category_box_add_items_button` name
